git-svn-id: https://unicode.org/repos/unicodetools/trunk@1381 13e8329f-0b23-4da4-9fe8-d0f6fe080806

+++ unicodetools/org/unicode/text/UCD/UCD.java
@@ -1042,7 +1042,7 @@ public final class UCD implements UCD_Types {
         return get(codePoint, false).joiningType;
     }
 
-    public byte getJoiningGroup(int codePoint) {
+    public short getJoiningGroup(int codePoint) {
         return get(codePoint, false).joiningGroup;
     }
